docs+ci: standards-conformance remediation pass (2026-07-05)
Corrects a false public-facing claim: the README asserted "all /STANDARDS
gates green" with no supporting table, when the audit found 33/138 controls
passing. README now has a real Standards Conformance table (11 rows) with
honest Applies/N-A states and gap-ledger links instead of the blanket claim.

Also fixes a real bug in Makefile's `security` target: the old
`tool && run || echo "skipping"` pattern didn't just soft-skip missing
tools, it also swallowed genuine tool failures — if gitleaks/bandit/
pip-audit were installed and found a real problem, the `||` still fired
and printed "skipping" over the failure, silently discarding it. `security`
now hard-fails on missing tools or real findings and is wired into
`make verify` for the first time. Hardening it surfaced 10 real pip-audit
CVEs in dev-toolchain transitive deps, all blocked on the repo's Python 3.9
floor; documented as dated waivers (Makefile) + a VEX
(docs/RESPONSIBLE-TECH-AUDITS.md) + fresh evidence in ADR-0002.

Other changes from the same pass: CI/Makefile parity (i18n gate, lint/type/
pwa-test routed through make targets), ruff findings fixed under an
expanded lint select + mccabe complexity budget, pytest strict-mode flags,
PEP 735 dependency-groups migration, derived __version__, Dockerfile digest
pin + HEALTHCHECK, CODEOWNERS + branch-ruleset definition (not yet applied
live), CHANGELOG/SECURITY/CONTRIBUTING/DEFINITION_OF_DONE, ADR log
(single-maintainer posture, Python-floor), a11y statement relocation, and
docs/GAP-LEDGER.md tracking every item not completed this pass.

Verified locally: 132/132 tests pass, ruff clean, mypy --strict clean,
`make verify` green end-to-end. Full narrative in
audit-2026-07-05/olive-bark-logger-REMEDIATION.md (Execution Log).

+++ report/charts.py
@@ -43,7 +43,7 @@ def bar_chart(
     bars: list[str] = []
     for i, (label, value) in enumerate(zip(labels, values)):
         x = gap + i * (bar_w + gap)
-        h = int(round((value / max_v) * chart_h)) if max_v else 0
+        h = round((value / max_v) * chart_h) if max_v else 0
         y = pad_top + (chart_h - h)
         bars.append(
             f'<rect x="{x}" y="{y}" width="{bar_w}" height="{h}" fill="{_BAR_COLOR}">'
